+++ spec/dummy/app/models/car.rb
@@ -1,0 +1,5 @@
+class Car < ActiveRecord::Base
+
+  include Storext
+
+end