+++ spec/dummy/db/migrate/20150113011502_create_cars.rb
@@ -1,0 +1,7 @@
+class CreateCars < ActiveRecord::Migration
+  def change
+    create_table :cars do |t|
+      t.hstore :data
+    end
+  end
+end